service/requirements.txt
# Core
kafka-python>=2.3.0
pycti>=6.0.0
python-dotenv>=1.0.0

# Feature extraction
scikit-learn>=1.4.0
numpy>=1.26.0

# RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
